refactor restore into storageStore
