+++ src/stores/mints.ts
@@ -507,18 +507,6 @@ export const useMintsStore = defineStore("mints", {
       }
       notifySuccess("Mint removed");
     },
-    restoreFromBackup: function (backup: any) {
-      if (!backup) {
-        notifyError("Unrecognized Backup Format!");
-      } else {
-        const keys = Object.keys(backup);
-        keys.forEach((key) => {
-          localStorage.setItem(key, backup[key]);
-        });
-        notifySuccess("Backup restored");
-        window.location.reload();
-      }
-    },
     assertMintError: function (response: { error?: any }, verbose = true) {
       if (response.error != null) {
         if (verbose) {
